pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eeg-emotional-state-classifier"
version = "0.1.0"
description = "Production-ready backend for EEG emotional state classification."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Project Contributors" }
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.1",
  "scipy>=1.11",
  "mne>=1.5",
  "scikit-learn>=1.3",
  "scikit-posthocs>=0.7.0",
  "xgboost>=1.7",
  "torch>=2.1",
  "joblib>=1.3",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.23",
  "typer>=0.9",
  "pydantic>=2.6",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5.0",
  "ipykernel"
]

[project.scripts]
eeg-emotion = "eeg_emotion.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["eeg_emotion*"]
